Allow GLES3 coexists with OpenGL on Linux
