+++ src/base/detect.h
@@ -51,7 +51,6 @@
 #define CONF_FAMILY_STRING "unix"
 #define CONF_PLATFORM_LINUX 1
 #define PLATFORM_STRING "linux"
-#define CONF_BACKEND_OPENGL_ES 1
 #define CONF_BACKEND_OPENGL_ES3 1
 #endif
