+++ src/engine/client/backend/glsl_shader_compiler.h
@@ -1,31 +1,13 @@
-#ifndef ENGINE_CLIENT_OPENGL_SL_H
-#define ENGINE_CLIENT_OPENGL_SL_H
-
-#include <base/detect.h>
-
-#include "graphics_defines.h"
+#ifndef ENGINE_CLIENT_BACKEND_GLSL_SHADER_COMPILER_H
+#define ENGINE_CLIENT_BACKEND_GLSL_SHADER_COMPILER_H
 
 #include <string>
 #include <vector>
 
-class CGLSLCompiler;
-
-class CGLSL
+enum EGLSLShaderCompilerType
 {
-public:
-	bool LoadShader(CGLSLCompiler *pCompiler, class IStorage *pStorage, const char *pFile, int Type);
-	void DeleteShader();
-
-	bool IsLoaded();
-	TWGLuint GetShaderID();
-
-	CGLSL();
-	virtual ~CGLSL();
-
-private:
-	TWGLuint m_ShaderID;
-	int m_Type;
-	bool m_IsLoaded;
+	GLSL_SHADER_COMPILER_TYPE_VERTEX = 0,
+	GLSL_SHADER_COMPILER_TYPE_FRAGMENT,
 };
 
 class CGLSLCompiler
@@ -65,7 +47,7 @@ public:
 	void AddDefine(const char *pDefineName, const char *pDefineValue);
 	void ClearDefines();
 
-	void ParseLine(std::string &Line, const char *pReadLine, int Type);
+	void ParseLine(std::string &Line, const char *pReadLine, EGLSLShaderCompilerType Type);
 
 	enum EGLSLCompilerTextureReplaceType
 	{
@@ -75,4 +57,4 @@ public:
 	};
 };
 
-#endif // ENGINE_CLIENT_OPENGL_SL_H
+#endif
