+++ src/engine/client/backend/opengl/opengl_sl_program.h
@@ -1,5 +1,14 @@
-#ifndef ENGINE_CLIENT_OPENGL_SL_PROGRAM_H
-#define ENGINE_CLIENT_OPENGL_SL_PROGRAM_H
+// This file can be included several times.
+#if(!defined(BACKEND_AS_OPENGL_ES) && !defined(ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H)) || \
+	(defined(BACKEND_AS_OPENGL_ES) && !defined(ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H_AS_ES))
+
+#if !defined(BACKEND_AS_OPENGL_ES) && !defined(ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H)
+#define ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H
+#endif
+
+#if defined(BACKEND_AS_OPENGL_ES) && !defined(ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H_AS_ES)
+#define ENGINE_CLIENT_BACKEND_OPENGL_OPENGL_SL_PROGRAM_H_AS_ES
+#endif
 
 #include <base/detect.h>
 
@@ -149,4 +158,4 @@ public:
 	int m_LocJumpIndex;
 };
 
-#endif // ENGINE_CLIENT_OPENGL_SL_PROGRAM_H
+#endif
