+++ src/engine/client/backend_sdl.h
@@ -7,10 +7,6 @@
 
 #include "graphics_defines.h"
 
-#ifdef CONF_BACKEND_OPENGL_ES3
-#define BACKEND_GL_MODERN_API 1
-#endif
-
 #include "blocklist_driver.h"
 #include "graphics_threaded.h"
 
@@ -111,342 +107,6 @@ struct SBackendCapabilites
 	int m_ContextPatch;
 };
 
-class CGLSLProgram;
-class CGLSLTWProgram;
-class CGLSLPrimitiveProgram;
-class CGLSLQuadProgram;
-class CGLSLTileProgram;
-class CGLSLTextProgram;
-class CGLSLPrimitiveExProgram;
-class CGLSLSpriteMultipleProgram;
-
-// takes care of opengl related rendering
-class CCommandProcessorFragment_OpenGL
-{
-protected:
-	struct CTexture
-	{
-		CTexture() :
-			m_Tex(0), m_Tex2DArray(0), m_Sampler(0), m_Sampler2DArray(0), m_LastWrapMode(CCommandBuffer::WRAP_REPEAT), m_MemSize(0), m_Width(0), m_Height(0), m_RescaleCount(0), m_ResizeWidth(0), m_ResizeHeight(0)
-		{
-		}
-
-		TWGLuint m_Tex;
-		TWGLuint m_Tex2DArray; //or 3D texture as fallback
-		TWGLuint m_Sampler;
-		TWGLuint m_Sampler2DArray; //or 3D texture as fallback
-		int m_LastWrapMode;
-
-		int m_MemSize;
-
-		int m_Width;
-		int m_Height;
-		int m_RescaleCount;
-		float m_ResizeWidth;
-		float m_ResizeHeight;
-	};
-	std::vector<CTexture> m_Textures;
-	std::atomic<int> *m_pTextureMemoryUsage;
-
-	TWGLint m_MaxTexSize;
-
-	bool m_Has2DArrayTextures;
-	bool m_Has2DArrayTexturesAsExtension;
-	TWGLenum m_2DArrayTarget;
-	bool m_Has3DTextures;
-	bool m_HasMipMaps;
-	bool m_HasNPOTTextures;
-
-	bool m_HasShaders;
-	int m_LastBlendMode; //avoid all possible opengl state changes
-	bool m_LastClipEnable;
-
-	int m_OpenGLTextureLodBIAS;
-
-	bool m_IsOpenGLES;
-
-public:
-	enum
-	{
-		CMD_INIT = CCommandBuffer::CMDGROUP_PLATFORM_OPENGL,
-		CMD_SHUTDOWN = CMD_INIT + 1,
-	};
-
-	struct SCommand_Init : public CCommandBuffer::SCommand
-	{
-		SCommand_Init() :
-			SCommand(CMD_INIT) {}
-		class IStorage *m_pStorage;
-		std::atomic<int> *m_pTextureMemoryUsage;
-		SBackendCapabilites *m_pCapabilities;
-		int *m_pInitError;
-
-		const char **m_pErrStringPtr;
-
-		char *m_pVendorString;
-		char *m_pVersionString;
-		char *m_pRendererString;
-
-		int m_RequestedMajor;
-		int m_RequestedMinor;
-		int m_RequestedPatch;
-
-		EBackendType m_RequestedBackend;
-
-		int m_GlewMajor;
-		int m_GlewMinor;
-		int m_GlewPatch;
-	};
-
-	struct SCommand_Shutdown : public CCommandBuffer::SCommand
-	{
-		SCommand_Shutdown() :
-			SCommand(CMD_SHUTDOWN) {}
-	};
-
-protected:
-	bool IsTexturedState(const CCommandBuffer::SState &State);
-	static bool Texture2DTo3D(void *pImageBuffer, int ImageWidth, int ImageHeight, int ImageColorChannelCount, int SplitCountWidth, int SplitCountHeight, void *pTarget3DImageData, int &Target3DImageWidth, int &Target3DImageHeight);
-
-	void InitOpenGL(const SCommand_Init *pCommand);
-
-	void SetState(const CCommandBuffer::SState &State, bool Use2DArrayTexture = false);
-	virtual bool IsNewApi() { return false; }
-	void DestroyTexture(int Slot);
-
-	static int TexFormatToOpenGLFormat(int TexFormat);
-	static int TexFormatToImageColorChannelCount(int TexFormat);
-	static void *Resize(int Width, int Height, int NewWidth, int NewHeight, int Format, const unsigned char *pData);
-
-	virtual void Cmd_Init(const SCommand_Init *pCommand);
-	virtual void Cmd_Shutdown(const SCommand_Shutdown *pCommand) {}
-	virtual void Cmd_Texture_Update(const CCommandBuffer::SCommand_Texture_Update *pCommand);
-	virtual void Cmd_Texture_Destroy(const CCommandBuffer::SCommand_Texture_Destroy *pCommand);
-	virtual void Cmd_Texture_Create(const CCommandBuffer::SCommand_Texture_Create *pCommand);
-	virtual void Cmd_Clear(const CCommandBuffer::SCommand_Clear *pCommand);
-	virtual void Cmd_Render(const CCommandBuffer::SCommand_Render *pCommand);
-	virtual void Cmd_RenderTex3D(const CCommandBuffer::SCommand_RenderTex3D *pCommand) {}
-	virtual void Cmd_Screenshot(const CCommandBuffer::SCommand_Screenshot *pCommand);
-
-	virtual void Cmd_Update_Viewport(const CCommandBuffer::SCommand_Update_Viewport *pCommand);
-	virtual void Cmd_Finish(const CCommandBuffer::SCommand_Finish *pCommand);
-
-	virtual void Cmd_CreateBufferObject(const CCommandBuffer::SCommand_CreateBufferObject *pCommand) {}
-	virtual void Cmd_RecreateBufferObject(const CCommandBuffer::SCommand_RecreateBufferObject *pCommand) {}
-	virtual void Cmd_UpdateBufferObject(const CCommandBuffer::SCommand_UpdateBufferObject *pCommand) {}
-	virtual void Cmd_CopyBufferObject(const CCommandBuffer::SCommand_CopyBufferObject *pCommand) {}
-	virtual void Cmd_DeleteBufferObject(const CCommandBuffer::SCommand_DeleteBufferObject *pCommand) {}
-
-	virtual void Cmd_CreateBufferContainer(const CCommandBuffer::SCommand_CreateBufferContainer *pCommand) {}
-	virtual void Cmd_UpdateBufferContainer(const CCommandBuffer::SCommand_UpdateBufferContainer *pCommand) {}
-	virtual void Cmd_DeleteBufferContainer(const CCommandBuffer::SCommand_DeleteBufferContainer *pCommand) {}
-	virtual void Cmd_IndicesRequiredNumNotify(const CCommandBuffer::SCommand_IndicesRequiredNumNotify *pCommand) {}
-
-	virtual void Cmd_RenderTileLayer(const CCommandBuffer::SCommand_RenderTileLayer *pCommand) {}
-	virtual void Cmd_RenderBorderTile(const CCommandBuffer::SCommand_RenderBorderTile *pCommand) {}
-	virtual void Cmd_RenderBorderTileLine(const CCommandBuffer::SCommand_RenderBorderTileLine *pCommand) {}
-	virtual void Cmd_RenderQuadLayer(const CCommandBuffer::SCommand_RenderQuadLayer *pCommand) {}
-	virtual void Cmd_RenderText(const CCommandBuffer::SCommand_RenderText *pCommand) {}
-	virtual void Cmd_RenderTextStream(const CCommandBuffer::SCommand_RenderTextStream *pCommand) {}
-	virtual void Cmd_RenderQuadContainer(const CCommandBuffer::SCommand_RenderQuadContainer *pCommand) {}
-	virtual void Cmd_RenderQuadContainerEx(const CCommandBuffer::SCommand_RenderQuadContainerEx *pCommand) {}
-	virtual void Cmd_RenderQuadContainerAsSpriteMultiple(const CCommandBuffer::SCommand_RenderQuadContainerAsSpriteMultiple *pCommand) {}
-
-public:
-	CCommandProcessorFragment_OpenGL();
-	virtual ~CCommandProcessorFragment_OpenGL() = default;
-
-	bool RunCommand(const CCommandBuffer::SCommand *pBaseCommand);
-};
-
-class CCommandProcessorFragment_OpenGL2 : public CCommandProcessorFragment_OpenGL
-{
-	struct SBufferContainer
-	{
-		SBufferContainer() {}
-		SBufferContainerInfo m_ContainerInfo;
-	};
-	std::vector<SBufferContainer> m_BufferContainers;
-
-	GL_SVertexTex3D m_aStreamVertices[1024 * 4];
-
-	struct SBufferObject
-	{
-		SBufferObject(TWGLuint BufferObjectID) :
-			m_BufferObjectID(BufferObjectID)
-		{
-			m_pData = NULL;
-			m_DataSize = 0;
-		}
-		TWGLuint m_BufferObjectID;
-		void *m_pData;
-		size_t m_DataSize;
-	};
-
-	std::vector<SBufferObject> m_BufferObjectIndices;
-
-#ifndef BACKEND_GL_MODERN_API
-	bool DoAnalyzeStep(size_t StepN, size_t CheckCount, size_t VerticesCount, uint8_t aFakeTexture[], size_t SingleImageSize);
-	bool IsTileMapAnalysisSucceeded();
-
-	void RenderBorderTileEmulation(SBufferContainer &BufferContainer, const CCommandBuffer::SState &State, const float *pColor, const char *pBuffOffset, unsigned int DrawNum, const float *pOffset, const float *pDir, int JumpIndex);
-	void RenderBorderTileLineEmulation(SBufferContainer &BufferContainer, const CCommandBuffer::SState &State, const float *pColor, const char *pBuffOffset, unsigned int IndexDrawNum, unsigned int DrawNum, const float *pOffset, const float *pDir);
-#endif
-
-	void UseProgram(CGLSLTWProgram *pProgram);
-
-protected:
-	void SetState(const CCommandBuffer::SState &State, CGLSLTWProgram *pProgram, bool Use2DArrayTextures = false);
-
-#ifndef BACKEND_GL_MODERN_API
-	void Cmd_Init(const SCommand_Init *pCommand) override;
-
-	void Cmd_RenderTex3D(const CCommandBuffer::SCommand_RenderTex3D *pCommand) override;
-
-	void Cmd_CreateBufferObject(const CCommandBuffer::SCommand_CreateBufferObject *pCommand) override;
-	void Cmd_RecreateBufferObject(const CCommandBuffer::SCommand_RecreateBufferObject *pCommand) override;
-	void Cmd_UpdateBufferObject(const CCommandBuffer::SCommand_UpdateBufferObject *pCommand) override;
-	void Cmd_CopyBufferObject(const CCommandBuffer::SCommand_CopyBufferObject *pCommand) override;
-	void Cmd_DeleteBufferObject(const CCommandBuffer::SCommand_DeleteBufferObject *pCommand) override;
-
-	void Cmd_CreateBufferContainer(const CCommandBuffer::SCommand_CreateBufferContainer *pCommand) override;
-	void Cmd_UpdateBufferContainer(const CCommandBuffer::SCommand_UpdateBufferContainer *pCommand) override;
-	void Cmd_DeleteBufferContainer(const CCommandBuffer::SCommand_DeleteBufferContainer *pCommand) override;
-	void Cmd_IndicesRequiredNumNotify(const CCommandBuffer::SCommand_IndicesRequiredNumNotify *pCommand) override;
-
-	void Cmd_RenderTileLayer(const CCommandBuffer::SCommand_RenderTileLayer *pCommand) override;
-	void Cmd_RenderBorderTile(const CCommandBuffer::SCommand_RenderBorderTile *pCommand) override;
-	void Cmd_RenderBorderTileLine(const CCommandBuffer::SCommand_RenderBorderTileLine *pCommand) override;
-#endif
-
-	CGLSLTileProgram *m_pTileProgram;
-	CGLSLTileProgram *m_pTileProgramTextured;
-	CGLSLPrimitiveProgram *m_pPrimitive3DProgram;
-	CGLSLPrimitiveProgram *m_pPrimitive3DProgramTextured;
-
-	bool m_UseMultipleTextureUnits;
-
-	TWGLint m_MaxTextureUnits;
-
-	struct STextureBound
-	{
-		int m_TextureSlot;
-		bool m_Is2DArray;
-	};
-	std::vector<STextureBound> m_TextureSlotBoundToUnit; //the texture index generated by loadtextureraw is stored in an index calculated by max texture units
-
-	bool IsAndUpdateTextureSlotBound(int IDX, int Slot, bool Is2DArray = false);
-
-public:
-	CCommandProcessorFragment_OpenGL2() :
-		CCommandProcessorFragment_OpenGL(), m_UseMultipleTextureUnits(false) {}
-};
-
-class CCommandProcessorFragment_OpenGL3 : public CCommandProcessorFragment_OpenGL2
-{
-};
-
-#define MAX_STREAM_BUFFER_COUNT 30
-
-// takes care of opengl 3.3+ related rendering
-class CCommandProcessorFragment_OpenGL3_3 : public CCommandProcessorFragment_OpenGL3
-{
-	bool m_UsePreinitializedVertexBuffer;
-
-	int m_MaxQuadsAtOnce;
-	static const int m_MaxQuadsPossible = 256;
-
-	CGLSLPrimitiveProgram *m_pPrimitiveProgram;
-	CGLSLPrimitiveProgram *m_pPrimitiveProgramTextured;
-	CGLSLTileProgram *m_pBorderTileProgram;
-	CGLSLTileProgram *m_pBorderTileProgramTextured;
-	CGLSLTileProgram *m_pBorderTileLineProgram;
-	CGLSLTileProgram *m_pBorderTileLineProgramTextured;
-	CGLSLQuadProgram *m_pQuadProgram;
-	CGLSLQuadProgram *m_pQuadProgramTextured;
-	CGLSLTextProgram *m_pTextProgram;
-	CGLSLPrimitiveExProgram *m_pPrimitiveExProgram;
-	CGLSLPrimitiveExProgram *m_pPrimitiveExProgramTextured;
-	CGLSLPrimitiveExProgram *m_pPrimitiveExProgramRotationless;
-	CGLSLPrimitiveExProgram *m_pPrimitiveExProgramTexturedRotationless;
-	CGLSLSpriteMultipleProgram *m_pSpriteProgramMultiple;
-
-	TWGLuint m_LastProgramID;
-
-	TWGLuint m_PrimitiveDrawVertexID[MAX_STREAM_BUFFER_COUNT];
-	TWGLuint m_PrimitiveDrawVertexIDTex3D;
-	TWGLuint m_PrimitiveDrawBufferID[MAX_STREAM_BUFFER_COUNT];
-	TWGLuint m_PrimitiveDrawBufferIDTex3D;
-
-	TWGLuint m_LastIndexBufferBound[MAX_STREAM_BUFFER_COUNT];
-
-	int m_LastStreamBuffer;
-
-	TWGLuint m_QuadDrawIndexBufferID;
-	unsigned int m_CurrentIndicesInBuffer;
-
-	void DestroyBufferContainer(int Index, bool DeleteBOs = true);
-
-	void AppendIndices(unsigned int NewIndicesCount);
-
-	struct SBufferContainer
-	{
-		SBufferContainer() :
-			m_VertArrayID(0), m_LastIndexBufferBound(0) {}
-		TWGLuint m_VertArrayID;
-		TWGLuint m_LastIndexBufferBound;
-		SBufferContainerInfo m_ContainerInfo;
-	};
-	std::vector<SBufferContainer> m_BufferContainers;
-
-	std::vector<TWGLuint> m_BufferObjectIndices;
-
-	CCommandBuffer::SColorf m_ClearColor;
-
-	void InitPrimExProgram(CGLSLPrimitiveExProgram *pProgram, class CGLSLCompiler *pCompiler, class IStorage *pStorage, bool Textured, bool Rotationless);
-
-protected:
-	static int TexFormatToNewOpenGLFormat(int TexFormat);
-	bool IsNewApi() override { return true; }
-
-	void UseProgram(CGLSLTWProgram *pProgram);
-	void UploadStreamBufferData(unsigned int PrimitiveType, const void *pVertices, size_t VertSize, unsigned int PrimitiveCount, bool AsTex3D = false);
-	void RenderText(const CCommandBuffer::SState &State, int DrawNum, int TextTextureIndex, int TextOutlineTextureIndex, int TextureSize, const float *pTextColor, const float *pTextOutlineColor);
-
-	void Cmd_Init(const SCommand_Init *pCommand) override;
-	void Cmd_Shutdown(const SCommand_Shutdown *pCommand) override;
-	void Cmd_Texture_Update(const CCommandBuffer::SCommand_Texture_Update *pCommand) override;
-	void Cmd_Texture_Destroy(const CCommandBuffer::SCommand_Texture_Destroy *pCommand) override;
-	void Cmd_Texture_Create(const CCommandBuffer::SCommand_Texture_Create *pCommand) override;
-	void Cmd_Clear(const CCommandBuffer::SCommand_Clear *pCommand) override;
-	void Cmd_Render(const CCommandBuffer::SCommand_Render *pCommand) override;
-	void Cmd_RenderTex3D(const CCommandBuffer::SCommand_RenderTex3D *pCommand) override;
-
-	void Cmd_CreateBufferObject(const CCommandBuffer::SCommand_CreateBufferObject *pCommand) override;
-	void Cmd_RecreateBufferObject(const CCommandBuffer::SCommand_RecreateBufferObject *pCommand) override;
-	void Cmd_UpdateBufferObject(const CCommandBuffer::SCommand_UpdateBufferObject *pCommand) override;
-	void Cmd_CopyBufferObject(const CCommandBuffer::SCommand_CopyBufferObject *pCommand) override;
-	void Cmd_DeleteBufferObject(const CCommandBuffer::SCommand_DeleteBufferObject *pCommand) override;
-
-	void Cmd_CreateBufferContainer(const CCommandBuffer::SCommand_CreateBufferContainer *pCommand) override;
-	void Cmd_UpdateBufferContainer(const CCommandBuffer::SCommand_UpdateBufferContainer *pCommand) override;
-	void Cmd_DeleteBufferContainer(const CCommandBuffer::SCommand_DeleteBufferContainer *pCommand) override;
-	void Cmd_IndicesRequiredNumNotify(const CCommandBuffer::SCommand_IndicesRequiredNumNotify *pCommand) override;
-
-	void Cmd_RenderTileLayer(const CCommandBuffer::SCommand_RenderTileLayer *pCommand) override;
-	void Cmd_RenderBorderTile(const CCommandBuffer::SCommand_RenderBorderTile *pCommand) override;
-	void Cmd_RenderBorderTileLine(const CCommandBuffer::SCommand_RenderBorderTileLine *pCommand) override;
-	void Cmd_RenderQuadLayer(const CCommandBuffer::SCommand_RenderQuadLayer *pCommand) override;
-	void Cmd_RenderText(const CCommandBuffer::SCommand_RenderText *pCommand) override;
-	void Cmd_RenderTextStream(const CCommandBuffer::SCommand_RenderTextStream *pCommand) override;
-	void Cmd_RenderQuadContainer(const CCommandBuffer::SCommand_RenderQuadContainer *pCommand) override;
-	void Cmd_RenderQuadContainerEx(const CCommandBuffer::SCommand_RenderQuadContainerEx *pCommand) override;
-	void Cmd_RenderQuadContainerAsSpriteMultiple(const CCommandBuffer::SCommand_RenderQuadContainerAsSpriteMultiple *pCommand) override;
-
-public:
-	CCommandProcessorFragment_OpenGL3_3() = default;
-};
-
 // takes care of sdl related commands
 class CCommandProcessorFragment_SDL
 {
@@ -488,10 +148,55 @@ public:
 	bool RunCommand(const CCommandBuffer::SCommand *pBaseCommand);
 };
 
+class CCommandProcessorFragment_OpenGLBase
+{
+public:
+	virtual ~CCommandProcessorFragment_OpenGLBase() = default;
+	virtual bool RunCommand(const CCommandBuffer::SCommand *pBaseCommand) = 0;
+
+	enum
+	{
+		CMD_INIT = CCommandBuffer::CMDGROUP_PLATFORM_OPENGL,
+		CMD_SHUTDOWN = CMD_INIT + 1,
+	};
+
+	struct SCommand_Init : public CCommandBuffer::SCommand
+	{
+		SCommand_Init() :
+			SCommand(CMD_INIT) {}
+		class IStorage *m_pStorage;
+		std::atomic<int> *m_pTextureMemoryUsage;
+		SBackendCapabilites *m_pCapabilities;
+		int *m_pInitError;
+
+		const char **m_pErrStringPtr;
+
+		char *m_pVendorString;
+		char *m_pVersionString;
+		char *m_pRendererString;
+
+		int m_RequestedMajor;
+		int m_RequestedMinor;
+		int m_RequestedPatch;
+
+		EBackendType m_RequestedBackend;
+
+		int m_GlewMajor;
+		int m_GlewMinor;
+		int m_GlewPatch;
+	};
+
+	struct SCommand_Shutdown : public CCommandBuffer::SCommand
+	{
+		SCommand_Shutdown() :
+			SCommand(CMD_SHUTDOWN) {}
+	};
+};
+
 // command processor impelementation, uses the fragments to combine into one processor
 class CCommandProcessor_SDL_OpenGL : public CGraphicsBackend_Threaded::ICommandProcessor
 {
-	CCommandProcessorFragment_OpenGL *m_pOpenGL;
+	CCommandProcessorFragment_OpenGLBase *m_pOpenGL;
 	CCommandProcessorFragment_SDL m_SDL;
 	CCommandProcessorFragment_General m_General;
